Reorganize field components
