--- libs/ui/src/components/checkbox-field/checkbox-field.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import CheckboxField from './checkbox-field';
-
-describe('Text', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<CheckboxField />);
-    expect(baseElement).toBeTruthy();
-  });
-});
