+++ libs/ui/src/components/fields/checkbox/checkbox.spec.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
 
-import PasswordField from '.';
+import CheckboxField from './checkbox';
 
 describe('Text', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<PasswordField />);
+    const { baseElement } = render(<CheckboxField />);
     expect(baseElement).toBeTruthy();
   });
 });
