+++ libs/ui/src/components/fields/field/helper/index.tsx
@@ -1,6 +1,6 @@
 import styles from './helper.module.css';
-import Text from '../../text';
-import { cx } from '../../../utils';
+import Text from '../../../text';
+import { cx } from '../../../../utils';
 
 export type FieldHelperType = 'helper' | 'error';
