+++ libs/ui/src/components/fields/password/strength-meter/index.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import { cx } from '../../utils';
-import styles from './password-strength-meter.module.css';
-import Field from '../field';
+import { cx } from '../../../../utils';
+import styles from './strength-meter.module.css';
+import Field from '../../field';
 import { useEffect, useState } from 'react';
 
 const conditions = [
