+++ libs/ui/src/components/fields/password/strength-meter/strength-meter.module.css
